Move Event back into root directory
Rather than modelling different kinds of events, it is more likely that
we will reuse the prop-types classes to model different event detail
classes.
