+++ src/event.ts
@@ -1,7 +1,24 @@
 import {
   ComponentCompilerEvent,
   ComponentCompilerMeta,
+  Config,
 } from '@stencil/core/internal';
+
+export function eventFromMetadata(
+  config: Config,
+  cmpMeta: ComponentCompilerMeta,
+  eventMeta: ComponentCompilerEvent,
+): Event {
+  const event = new Event(cmpMeta, eventMeta);
+
+  if (!event.isSupported()) {
+    config.logger?.warn(
+      `Component "${cmpMeta.tagName}" event "${eventMeta.name}" is not supported by Elm output target.`,
+    );
+  }
+
+  return event;
+}
 
 export class Event {
   // TODO support decoding CustomEvent detail values
